Extract MaxRequests and Oom to separate source files
