+++ lib/unicorn/worker_killer/configuration.rb
@@ -1,11 +1,13 @@
-module Unicorn::WorkerKiller
-  class Configuration
-    attr_accessor :max_quit, :max_term, :sleep_interval
+module Unicorn
+  module WorkerKiller
+    class Configuration
+      attr_accessor :max_quit, :max_term, :sleep_interval
 
-    def initialize
-      self.max_quit = 10
-      self.max_term = 15
-      self.sleep_interval = 1
+      def initialize
+        self.max_quit       = 10
+        self.max_term       = 15
+        self.sleep_interval = 1
+      end
     end
   end
 end
